Added simple user validator

+++ routes/UsersRoutes.ts
@@ -1,14 +1,17 @@
 import { CommonRoutesConfig } from './CommonRoutesConfig';
 import express from 'express';
 import { LocalUsers } from '../repository/users/LocalUsers';
-import { IUserModel } from "../models/IUserModel";
+import { IUserModel } from '../models/IUserModel';
 import { RepoBase } from '../repository/IRepository';
+import { UserValidator} from '../validators/UserValidator';
 
 export class UsersRoutes extends CommonRoutesConfig {
     repo: RepoBase<IUserModel, string>;
+    validator: UserValidator;
     constructor(app: express.Application) {
         super(app, 'UsersRoutes');
         this.repo = new LocalUsers();
+        this.validator = new UserValidator();
     }
 
     configureRoutes() {
@@ -18,6 +21,9 @@ export class UsersRoutes extends CommonRoutesConfig {
             })
             .post((req: express.Request, res: express.Response) => {
                 const user: IUserModel = req.body;
+                if(!this.validator.Validate(user)){
+                    return res.status(400).send(this.validator.GetMessage(user));
+                }
                 res.status(200).send(this.repo.create(user));
             });
 
@@ -30,10 +36,16 @@ export class UsersRoutes extends CommonRoutesConfig {
             })
             .put((req: express.Request, res: express.Response) => {
                 const user: IUserModel = req.body;
+                if(!this.validator.Validate(user)){
+                    return res.status(400).send(this.validator.GetMessage(user));
+                }
                 res.status(200).send(this.repo.update(user));
             })
             .delete((req: express.Request, res: express.Response) => {
                 const user: IUserModel = req.body;
+                if(!this.validator.Validate(user)){
+                    return res.status(400).send(this.validator.GetMessage(user));
+                }
                 res.status(200).send(this.repo.delete(user));
             });
